[Modernization] Addressed PR feedback

+++ srecord/arglex/tool.cc
@@ -296,7 +296,7 @@ srecord::arglex_tool::get_address_and_nbytes(const char *name,
     {
         nbytes = get_number("byte count", 1, 8);
     }
-    if ((long long)address + nbytes > (1LL << 32))
+    if ((uint64_t)address + nbytes > (1LL << 32))
     {
         fatal_error
         (
@@ -326,7 +326,7 @@ srecord::arglex_tool::get_address_nbytes_width(const char *name,
             width = get_number("width", 1, nbytes);
         }
     }
-    if ((long long)address + nbytes > (1LL << 32))
+    if ((uint64_t)address + nbytes > (1LL << 32))
     {
         fatal_error
         (
